fix : 화살표 아이콘 클릭시 컨텐츠만 렌더링 될 수 있도록 컴포넌트를 분리한다
