+++ src/Router.tsx
@@ -1,6 +1,6 @@
 import { createBrowserRouter } from "react-router-dom";
 import Layout from "./pages/Layout";
-import Home from "./pages/Home";
+import Home from "./pages/home/Home";
 
 export const router = createBrowserRouter([
   {
